% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificates.R
\docType{class}
\name{certificates}
\alias{certificates}
\alias{certs}
\title{Certificates in Key Vault}
\description{
This class represents the collection of certificates stored in a vault. It provides methods for managing certificates, including creating, importing and deleting certificates, and doing backups and restores. For operations with a specific certificate, see \link{certificate}.
}
\section{Methods}{

This class provides the following methods:\preformatted{create(name, subject, x509=cert_x509_properties(), issuer=cert_issuer_properties(),
       key=cert_key_properties(), format=c("pem", "pkcs12"),
       actions=cert_expiry_actions(),
       attributes=vault_object_attrs(),
       ..., wait=TRUE)
import(name, value, pwd=NULL,
       attributes=vault_object_attrs(),
       ..., wait=TRUE)
get(name)
delete(name, confirm=TRUE)
list()
backup(name)
restore(backup)
get_contacts()
set_contacts(email, name, phone)
}
}

\section{Arguments}{

\itemize{
\item \code{name}: The name of the certificate.
\item \code{subject}: For \code{create}, The subject or X.500 distinguished name for the certificate.
\item \code{x509}: Other X.509 properties for the certificate, such as the domain name(s) and validity period. A convenient way to provide this is via the \link{cert_x509_properties} helper function.
\item \code{issuer}: Issuer properties for the certificate. A convenient way to provide this is via the \link{cert_issuer_properties} helper function. The default is to specify a self-signed certificate.
\item \code{key}: Key properties for the certificate. A convenient way to provide this is via the \link{cert_key_properties} helper function.
\item \code{format}: The format to store the certificate in. Can be either PEM or PFX, aka PKCS#12. This also determines the format in which the certificate will be exported (see \link{certificate}).
\item \code{actions}: The actions to take when the certificate is about to expire. A convenient way to provide this is via the \link{cert_expiry_actions} helper function.
\item \code{attributes}: Optional attributes for the secret. A convenient way to provide this is via the \link{vault_object_attrs} helper function.
\item \code{value}: For \code{import}, the certificate to import. This can be the name of a PFX file, or a raw vector with the contents of the file.
\item \code{pwd}: For \code{import}, the password if the imported certificate is password-protected.
\item \code{...}: For \code{create} and \code{import}, other named arguments which will be treated as tags.
\item \code{wait}: For \code{create} and \code{import}, whether to wait until the certificate has been created before returning. If FALSE, you can check on the status of the certificate via the returned object's \code{sync} method.
\item \code{backup}: For \code{restore}, a string representing the backup blob for a key.
\item \code{email}: For \code{set_contacts}, the email addresses of the contacts.
}
}

\section{Value}{

For \code{get}, \code{create} and \code{import}, an object of class \code{stored_certificate}, representing the certificate itself.

For \code{list}, a vector of key names.

For \code{backup}, a string representing the backup blob for a certificate. If the certificate has multiple versions, the blob will contain all versions.
}

\examples{
\dontrun{

vault <- key_vault$new("mykeyvault")

vault$certificates$create("mynewcert", wait=TRUE)
vault$certificates$list()
vault$certificates$get("mynewcert")

# specifying a validity period of 2 years (24 months)
vault$certificates$create("mynewcert", x509=cert_x509_properties(valid=24))

# setting management tags
vault$certificates$create("mynewcert", tag1="a value", othertag="another value")

# importing a cert from a PFX file
vault$certificates$import("importedcert", "mycert.pfx")

# backup and restore a cert
bak <- vault$certificates$backup("mynewcert")
vault$certificates$delete("mynewcert", confirm=FALSE)
vault$certificates$restore(bak)

# set a contact
vault$certificates$set_contacts("username@mydomain.com")
vault$certificates$get_contacts()

}
}
\seealso{
\link{certificiate}, \link{cert_key_properties}, \link{cert_x509_properties}, \link{cert_issuer_properties}, \link{vault_object_attrs}

\href{https://docs.microsoft.com/en-us/azure/key-vault/}{Azure Key Vault documentation},
\href{https://docs.microsoft.com/en-us/rest/api/keyvault}{Azure Key Vault API reference}
}
