% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{list_deleted_key_vaults}
\alias{list_deleted_key_vaults}
\title{List soft-deleted Key Vaults}
\description{
Method for the \link[AzureRMR:az_subscription]{AzureRMR::az_subscription} class.
}
\section{Usage}{
\preformatted{list_deleted_key_vaults()
}
}

\section{Value}{

This method returns a data frame with the following columns:
\itemize{
\item \code{name}: The name of the deleted key vault.
\item \code{location}: The location (region) of the vault.
\item \code{deletion_date}: When the vault was soft-deleted.
\item \code{purge_date}: When the vault is scheduled to be purged (permanently deleted).
\item \code{protected}: Whether the vault has purge protection enabled. If TRUE, manual attempts to purge it will fail.
}
}

\seealso{
\link{create_key_vault}, \link{get_key_vault}, \link{delete_key_vault}, \link{purge_key_vault}, \link{az_key_vault},

\href{https://docs.microsoft.com/en-us/azure/key-vault/}{Azure Key Vault documentation},
\href{https://docs.microsoft.com/en-us/rest/api/keyvault}{Azure Key Vault API reference}
}
