% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_endpoint.R
\docType{class}
\name{key_vault}
\alias{key_vault}
\title{Azure Key Vault endpoint class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
key_vault
}
\description{
Class representing the client endpoint for a key vault, exposing methods for working with it.
}
\section{Fields}{

\itemize{
\item \code{keys}: A sub-object for working with encryption keys stored in the vault. See \link{keys}.
\item \code{secrets}: A sub-object for working with secrets stored in the vault. See \link{secrets}.
\item \code{certificates}: A sub-object for working with certificates stored in the vault. See \link{certificates}.
\item \code{storage}: A sub-object for working with storage accounts managed by the vault. See \link{storage}.
}
}

\section{Methods}{

This class provides one method, for initialization:\preformatted{new(url, tenant = "common", app = .az_cli_app_id, ...,
    domain = "vault.azure.net", token = NULL)
}

The arguments are as follows:
\itemize{
\item \code{url}: The location of the vault. This can be a full URL, or the vault name alone; in the latter case, the \code{domain} argument is appended to obtain the URL.
\item \code{tenant, app, ...}: Authentication arguments that will be passed to \link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token}. The default is to authenticate interactively.
\item \code{domain}: The domain of the vault; for the public Azure cloud, this is \code{vault.azure.net}. Also the resource for OAuth authentication.
\item \code{token}: An OAuth token obtained via \link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token}. If provided, this overrides the other authentication arguments.
}

To work with objects stored in the key vault, use the methods provided by one of the sub-objects listed in 'Fields'.
}

\examples{
\dontrun{

key_vault$new("mykeyvault")
key_vault$new("https://mykeyvault.vault.azure.net")

# authenticating as a service principal
key_vault$new("mykeyvault", tenant="myaadtenant", app="app_id", password="password")

# authenticating with an existing token
token <- AzureAuth::get_azure_token("https://vault.azure.net", "myaadtenant",
                                    app="app_id", password="password")
key_vault$new("mykeyvault", token=token)

}
}
\seealso{
\link{az_key_vault}, \link{keys}, \link{secrets}, \link{certificates}, \link{storage}

\href{https://docs.microsoft.com/en-us/azure/key-vault/}{Azure Key Vault documentation},
\href{https://docs.microsoft.com/en-us/rest/api/keyvault}{Azure Key Vault API reference}
}
\keyword{datasets}
